Fix update of tool icon and cursor hot spot
The signal for changing the icon of statusbar and cursor was
buttonPressed. The user could press and drag a tool button and change
the icon and cursor without changing the tool for drawing.
Now the signal buttonToggled solves just that.
Besides, the cursor hot spot is now defined as the bottom left point of
the pixmap used.