Fix ghost items
The items drawn were being deleted from scene after 'new' or 'open' menus
were triggered. But then they would appear again when choosing the same tool
to begin drawing in the new file.
Now the tools are being reset when these menus are triggered, and a new function
was created in canvas.py for that.