Modify canvas dimensions and resolve scaling issues
Canvas had 640x460 dimension, accounting for a ration of ~1,39, while
the ratio of steps measured is ~1,1633, i.e. HEIGHT_STEP = ~0,86*WIDTH_STEP.
It hampered a reasonable scaling and the use of all plotter area.
Now canvas has 555x477, accounting for a ratio of ~1,1635.
In addition, every point (x, y) in canvas, will be scaled as follows:
    - x_to_plot = x . (WIDTH_STEP/CANVAS_WIDTH)
    - y_to_plot = y . (WIDTH_STEP/CANVAS_WIDTH)
so that for the point (0, 0) in GUI, the plotter will receive (0, 0) and
for the point 555,477, the plotter will receive (WIDTH_STEP, ~0,86*WIDTH_STEP).